{EQEmu PR 3351} [Bug fix] fix for mob scaling issue with min dmg set to zero while max dmg is not.
A user on Ret submitted a bug where they were one shot by a low level mob. I was able to reproduce this and started to diagnose the issue. It seemed to be where a Min Dmg value of 0 and a max dmg value of 3 was set in the database for this NPC.
![image](https://github.com/EQEmu/Server/assets/4657161/5362fa99-70e2-4cb9-948f-9acf1c3c9563)

The min dmg scaling kicks in, but you get into a situation where MIN DMG > MAX DMG (12 vs 3 in this case)
It leads to situations like thus.
![image](https://github.com/EQEmu/Server/assets/4657161/cf15912f-e133-40a5-b2c0-d3a7630edcb4)

Looking at it deeper it seems to be in the calculation area for rounding.

![image](https://github.com/EQEmu/Server/assets/4657161/c31afa87-1d13-4d25-a240-47ad9fa6030a)

I worked with Kingly and setup some sanity code to prevent wild situations of being hit for millions of damage, but needed to go one step further.

I also changed it so that if you set Min Dmg to 0, it means max damage is also auto calculated... this had the mob working as I would expect a lvl 28 mob to be.

![image](https://github.com/EQEmu/Server/assets/4657161/e644418a-216f-4ba9-a233-0580282fdf91)