{EQEmu PR 3383} [Targeting] /tar <bad target> should not untarget existing target
On live if you have something targetted, and you type /tar Innothule, but Innothule is not around, you get an error message and your current target is unmodifed.

Prior to this patch, if you do this on eqemu, it "looks" like your current target is unchanged, but you no longer get updates.

This change fixes that.  But I have a couple of questions:

1. The code differentiates between AI controlled clients and non-AI controlled clients.  Is not the /tar command unable to be send if you are AI controlled, meaning I could also delete those lines?  If not, perhaps the code I moved down inside non-ai controlled needs to be down in AI controlled as well?  My take is that those lines can be deleted.
2. The ONLY time all of the code after the // HoTT comment executes if its a NON-AI-CONTROLLED successful target.  So I don't believe any of that needs to be changed, since that's the case I fixed.  Just looking for a second set of eyes on that.