{EQEmu PR 3407} [Targetting] Fix bug when using /tar on invalid target
Using /tar on an invalid target should not negate a current target. (Tested on live)

Mt previous attempt to fix this was invalid, as Handle_OP_Target was getting called for multiple packets, so my fix could not reside there alone.

In my digging I found that hitting escape (to remove target), mobs dying, clicking with the mouse, all issued direct OP_TargetMouse commands which should bahave as we have always done them.  However, our code also called the same code for OP_Target, which actually only seems to be meant to server as a validation of a /tar command.  If we accept the target, the client immediately sends the corresponding OP_TargetMouse.  If we reject it, the client sends nothing further.

This fix seems to solve the original issue while not breaking the rest, which is much better than the last shitty PR I submitted.

The diff is hard to read because of indentation changes caused by code removal.  Handle_TargetMouse is exactly whaty Handle_Target used to be, except it mo longer handes accept/reject.  HandleTarget has been refactored for clarity, and only handles accept/reject.