{EQEmu PR 3458} [Bug Fix] Add check for underscores in botcreate command
### NOTES
Check for underscores in bot names as the server will replace with a space and add to the database. The bot can't be deleted/summoned or changed except by editing the database.

Other symbols seem to be properly ignored (they are just deleted, like # or *). I added a simple check to prevent names with _'s.

### Steps to Recreate
**Command:** ^botcreate Tuday_bot_healer 1 1 1

![image](https://github.com/EQEmu/Server/assets/27604078/76a9f690-e15e-4ac9-b5e3-eda59a54ff3f)

The server replaces _'s with spaces and adds the bot. Here is an example:

![image](https://github.com/EQEmu/Server/assets/27604078/881b3a42-9357-4792-96b8-15d6e32bc044)

You can no longer delete/summon/alter the bot in anyway without editing SQL.

### Example after adding check

![image](https://github.com/EQEmu/Server/assets/27604078/1831a0c4-7970-411a-bd28-4496677069c6)