{EQEmu PR 3533} [Rules] Add rule to toggle pets accepting quest items
Adds the following rule: Pets:CanTakeQuestItems, default true.

As name suggests, rejects items flagged as "Quest" when traded to a pet. The main purpose for this rule is to block players accidentally trading quest items to their pet instead of the desired quest NPC.

Everything works so far in testing.