{EQEmu PR 3549} [Charm] Fix to update target windows on charm on/off
On classic servers that do not use the rule AlwaysSendTargetsBuffs, we can only see a charmed pet's buffs if we manually retarget after the charm.  These buffs also remain displayed (including the charm) after charm breaks unless we re-target.  This is a bug.

This PR fixes this issue by updating all client target windows when a mob they have targeted is charmed.  It also clears all the buffs from that clients sight when the charm drops (unless client is a GM, AlwaysSendTargetsBuffs is set, or the groupAAInspectBuffsis in play).