{EQEmu PR 3563} [Instances] Honor reserved instances
Logic to select next available instance id was incorrect. The correct logic selected the max id + 1 or max reserved + 1, but then it would overwrite if we enabled recycling ids. Additionally, it was incrementing the reserved ids and assigning ids to the max reserved id vice the next available. Finally, it was running the logic twice.