{EQEmu PR 3577} [Commands] npc_edit faction and #setfaction duplicate and incorrect.
The #npcedit faction <#> command replaces the value of npc_faction_id in npc_types.  This is NOT a faction_id.  This is a reference into the npc_faction table.  That table then assigns a primary_faction (which is a faction ID) and also references any hits or assist preferences in the npc_faction_entries table.

The old code for #npcedit faction command would set the npc_faction_id and then  lie and lookup a faction in faction_list and show that name as the npc's new faction.  That is totally random.

This new code deferences the value provided, and if it is a valid npc_faction_id, uses that npc_faction's primary faction to display what the npc is now on.

It should be noted, this in game command is a hack.  You need access to the npc_faction list before you can use this, at which point you're probably better suited not to make this change in game.

That said, it exists, so I made it actually report accurately what it was doing.

The #setfaction command was doing the same thing, in the same incorrect way.  I don't think it should exist.