{EQEmu PR 3600} [Bug Fix] Add check to prevent npcs summoning across zones
This is an attempt at fixing #1523

Edit: Testing complete. Through exhaustive testing, I can confirm that adding this check does not break anything in this functionality. I cannot however 100% confirm that it resolves the underlying bug as I cannot seem to reproduce it. My recommendation is to merge this and see if it does the trick for those that are experiencing the original bug. Worst case scenario, we are no worse off than we were prior to this check being added if it misses the mark on fixing the bug.