{EQEmu PR 3664} [Spawn] Split spawn2 enabled into its own state table
**Problem**

Currently `spawn2` is classified as a content table. A content table should be read-only from a server mutability perspective. Meaning the only thing that should be changing content tables should be a developer. Server state should be separately maintained.

On PEQ the content tables are on a read only CDN so we get errors like

```
[QueryErr] Zone [sharvahl] [1146] [Table 'peq.spawn2' doesn't exist]
[UPDATE spawn2 SET enabled = 1 WHERE id = 32803]
```

Because `spawn2` **enabled** field database calls are set to a non-content connection while `spawn2` is classified as a content table on the content connection

**Solution**

Split out the **enabled** field from `spawn2` onto its own table called `spawn2_disabled` which inversely reads from the prior implementation of **enabled** but it is so we are not storing all `spawn2` entries as **enabled** in another table. We only need to know about the ones that are disabled otherwise we assume all spawns are enabled.

**Additional Notes**

This also implements enabled/disabled to be instance aware and @fryguy503 (Trust) owes me a nice box of pie.