{EQEmu PR 3686} [Illusions] RandomizeFeatures erased texture.
This PR re-fixes the issue I fixed in this [PR](https://github.com/EQEmu/Server/pull/3376).

It was rebroken as a result of many fixes put in via this [PR](https://github.com/EQEmu/Server/pull/3520).

The problem is that when using RandomizeFeatures (which does not alter texure), textture gets set back to 0xFF.

I reapplied my fix to the new code and it once again works properly.

To see bug: spawn an npc wth a custom texture (like 12).
Use RandomizeFeatures.
Texture is lost.

This fixes it by not saving the texture if it was not supplied as a value other that 0xFF.