{EQEmu PR 3705} [Bug Fix] Add "IgnoreLevelBasedHasteCaps" rule to GetHaste() to fix showstats
I PR'd this rule a while back but only modified the CalcHaste function in client_mods.cpp. It wasn't until recently I discovered that #mystats/#showstats was displaying incorrect haste values. After making a thread in discord about it Nite pointed out that showstats uses GetHaste in mob.cpp to calculate haste to display and suggested we should probably update GetHaste to incorporate this rule. This seems to work as expected and I'm getting proper haste display in #showstats as of now. I made sure to include IsClient() checks as Nite suggested since this function is also used by mobs.