{EQEmu PR 3718} [Faction] Alliance line is only allowed 1 faction change at a time.
The alliance line of faction spells can only impact (1) faction at a time.  The existing code allows an infinate #.

This video shows on live how it works:

https://github.com/EQEmu/Server/assets/8644833/9977a8ba-caf8-4e79-bff4-28ffeb874ad1

Looking at my local DB and projecteq, there are only 6 spells that use SE_AddFaction and they are all the alliance line, or a guide version of it.

If someone has added a custom SE_AddFaction to there server, and wants it stick, regardless of the alliance line, we'd need something further.  But as it stands, this is much siumpler, and more accurate.