{EQEmu PR 3811} [Combat] Disarm was not dropping item to ground due to bug
The disarm ability should remove the equipped weapon and drop it to the ground so long as it is non-magical and not no-drop.

The code was testing no-drop against the value 255.

Most of the db I have seen use the values of 0 (not droppable) and 1 (droppable).

This caused those servers (afaik all servers) to not do this part of the code correctly.