{EQEmu PR 4238} [Bug Fix] SPA214 SE_MaxHPChange calculation errors corrected.
Bug: SPA 214 which changes hit points by percentage was not displaying correctly on the client. This was causing differences between client displayed hit points and server calculations.

Solution: The formula we were using to apply the spell and item bonuses using SPA 214 to client max hit points was incorrect. It should not have been applying it after the AA version (Natural Durability) and after Spell HP buffs (Ie. Virtue). The correct calculation was it should be applied at the same time as the AA version.

Client and server now display the same values within 5 hps of each other.

Fixed for Bots and Mercs as well.